+++ targets/TARGET_STM/TARGET_STM32F4/common_objects.h
@@ -110,6 +110,7 @@ struct i2c_s {
 #endif
 };
 
+#define GPIO_IP_WITHOUT_BRR
 #include "gpio_object.h"
 
 #ifdef __cplusplus
