+++ targets/TARGET_STM/gpio_api.c
@@ -34,19 +34,19 @@
 
 extern uint32_t Set_GPIO_Clock(uint32_t port_idx);
 
-uint32_t gpio_set(PinName pin)
-{
+uint32_t gpio_set(PinName pin) {
     MBED_ASSERT(pin != (PinName)NC);
 
     pin_function(pin, STM_PIN_DATA(STM_MODE_INPUT, GPIO_NOPULL, 0));
+
     return (uint32_t)(1 << ((uint32_t)pin & 0xF)); // Return the pin mask
 }
 
-void gpio_init(gpio_t *obj, PinName pin)
-{
+void gpio_init(gpio_t *obj, PinName pin) {
     obj->pin = pin;
-    if (pin == (PinName)NC)
+    if (pin == (PinName)NC) {
         return;
+    }
 
     uint32_t port_index = STM_PORT(pin);
 
@@ -58,16 +58,18 @@ void gpio_init(gpio_t *obj, PinName pin)
     obj->mask    = gpio_set(pin);
     obj->reg_in  = &gpio->IDR;
     obj->reg_set = &gpio->BSRR;
+#ifdef GPIO_IP_WITHOUT_BRR
+    obj->reg_clr = &gpio->BSRR;
+#else
     obj->reg_clr = &gpio->BRR;
+#endif
 }
 
-void gpio_mode(gpio_t *obj, PinMode mode)
-{
+void gpio_mode(gpio_t *obj, PinMode mode) {
     pin_mode(obj->pin, mode);
 }
 
-void gpio_dir(gpio_t *obj, PinDirection direction)
-{
+void gpio_dir(gpio_t *obj, PinDirection direction) {
     MBED_ASSERT(obj->pin != (PinName)NC);
     if (direction == PIN_OUTPUT) {
         pin_function(obj->pin, STM_PIN_DATA(STM_MODE_OUTPUT_PP, GPIO_NOPULL, 0));
